Add comments & rename composeIndex -> composeDataIndex
